Handle skipped includes in inclusion rewriter
If a header that ought to be rewritten is included more than once (e.g. through
transitive inclusions), the previous behavior would not comment out any
additional #include directives after the first one, due to how inclusions are
tracked by the InclusionRewriter.

In the Clang mainline version of InclusionRewriter this is not an issue, as
every directive is commented out. Since hipSYCL only rewrites certain
inclusions, this becomes a problem.

This change adds an additional data structure that keeps track of any skipped
inclusions, as well as a check whether an #include directive is for a skipped
inclusion and should thus be commented out.