Add support for compiling CUDA using Clang
This adds a new method for compiling CUDA - using Clang - which replaces nvcc
as the default compiler for CUDA. The previous behavior can be restored by
passing "--hipsycl-platform=nvcc".

The version of Clang used for compilation can be configured using either the
"--cuda-clang-compiler=<compiler>" CLI parameter, or the
HIPSYCL_CUDA_CLANG_COMPILER environment variable.

In addition, this includes some minor refactorings of some parts of syclcc and
adds a new configuration option "platform_needs_template_pruning", which
dictates whether the "--with-template-pruning" flag is passed to
hipsycl_transform_source. This is only enabled for nvcc.