Fix item id calculations & improve spec conformance
This fixes the way work items are selected/discarded before the
user-provided kernel functor is called. Previously this was based on the
linear id of a work item, which causes problems whenever the global size
is not a multiple of the default work group size (e.g. 16/16 for 2D
kernels).

Furthermore detail::item_impl now receives the global size of the kernel
instead of computing it based on the HIP builtins such as hipBlockDim_*,
as those again always are a multiple of the default work group size,
which may not coincide with the user-defined global size. This would
then cause wrong results in item::get_linear_id() and importantly also
for the linear buffer offsets computed within accessors.

For the item variant containing an offset, the subscript operator[] now
correctly adds the offset to the returned index (like it was already
done in item::get_id()).

Finally, the item API is updated to conform to SYCL 1.2.1 revision 4.

Tests are provided for all changes.