Switch to out-of-source builds
This changes the location of temporary files generated by the inclusion
rewriter and transformation passes from the source file location to a
unique temporary directory for each translation unit. As the original
causes for why in-source builds where preferred are no longer relevant
(see #21), this constitutes an overall cleaner and safer approach.
Furthermore, for certain operations, for example CMake's try_compile, it
is imperative that the original source directory remains untouched, as
it might not be writable by the user (thus breaking CMake configuration).