Refactor item implementation
This refactors the implementation of cl::sycl::item to use partial
template specializations over SFINAE. Instead of relying on a
item_impl that is inaccessible to users, detail::make_item() can now be
used to instantiate items.

The primary reason for this change is to silence a faulty Clang
diagnostic about the conversion operator from item<dims, false> to
item<dims, true> never being used. However, as an added benefit, the new
implementation is also somewhat cleaner, as it more clearly separates
the differences between items with and without offsets, while still
maintaining a zero-overhead implementation for the latter.