[CMake] Handle RPATH, explicitly link libhipSYCL
This improves the SYCL CMake language by adding proper support for RPATH
flags for compiling and linking. Furthermore, hipSYCL now also exports a
list of all targets (i.e., libhipSYCL_cpu and so on) which can be used
to link to it directly from within CMake. While syclcc already takes
care of this in general, explicitly linking to a libhipSYCL_* library
allows CMake to write the proper RPATH into the resulting executable.

This also ensures that any target on which add_sycl_to_target() is
called is always linked using syclcc (even if it contains some non-SYCL
translation units), which is required as certain additional libraries
(e.g. the CUDA runtime) are only added by syclcc.