clang plugin: Fix incorrect function stripping from IR if unknown functions appear in the IR
If functions appear in the IR whose mangled name we have not seen in the IR, those functions would currently not be stripped.
This is because we assume that everything that all functions that are strippable are marked as implicitly __host__ __device__
during the AST pass.
This commit fixes the issue by reversing the logic:
We remember functions that were explicitly __global__, __sycl_kernel or __device__ during the AST pass, and only prevent stripping
of functions from the IR if they had these attributes attached.