clang plugin: Fix variables not being correctly allocated in shared memory in hierarchical parallel for
In order to have a variable in __shared__ memory, it is not sufficient to just add the __shared__ attribute. It is also required to set the storage class of the variable to static.
Previously, the latter was missing which caused variables in parallel hierarchical for invocations not to be put properly in __shared__ memory, even though the attribute was correctly added.
This commit fixes that.