clang plugin: Fix IR function pruning for cycles in call graph
The previous method of determining whether a function can be pruned from the IR relied on determining "backwards" if one of the callers of the function is an entrypoint.
This approach is prone to infinite loops when there are cycles in the callgraph.

This commit fixes this by creating an explicit callgraph, finding all children of the entrypoints in the callgraph, and removing functions if they are not among the children of entrypoints.