clang plugin: Fix IR pruning edge cases, refactor
This fixes two edge cases related to IR device function pruning that
were encountered when trying to compile a larger codebase.

First, the pruning is now performed in an LLVM ModulePass, instead of a
FunctionPass. This is because the latter's runOnFunction callback
seemingly does not always receive all functions (possibly related to
delayed instantiation of templates?) which in turn leads to those
functions not being pruned, causing ptxas errors.

The second issue is more obscure and the root cause is hard to pinpoint.
Symptoms again include various ptxas errors, and the problem appears to
be related to non-device code (e.g. exceptions) being included in the IR
through libraries with CUDA support that contain explicit __device__
functions (e.g. Boost). As a workaround, the pruning pass now only
considers kernels (i.e., functions marked as __global__) as entrypoints.
Previously any function explicitly marked as __device__ was considered
as well. This change should not be a problem, as Clang currently does
not support device object linking.

This also refactors the IR pruning pass by replacing the custom
CallGraph with LLVM's implementation. The main difference between the
two is that the latter does not include nodes for LLVM intrinsics. For
this reason, we simply don't prune any intrinsics at all.