Small improvements on cmake compiler and library discover:
There are mainly three changes regarding the build file:

a) changed the libhipSYCL compiler to use CMAKE_CXX_COMPILER instead
of default cc/cxx by supplying a --force-alternative-compiler option.
In some systems, the default cc/cxx points to an ancient compiler, and
a regular user has no control over it. If the syclcc script happens to find
these things, the compiling of the software will choke.

b) change reference of -lLLVM to ${LLVM_LIBS}, as the latter is discovered
by find_package(LLVM), the former is not. As a result, if the llvm libraries
are not in the default path, the linker will not be able to find the llvm libraries.

c) by similar rationale, link to the target exported by find_pacakge(Clang)
instead of explicit specify -lclangAST etc.