[CMake] Enable dependency file output for SYCL
This copies the CMAKE_DEPFILE_FLAGS_ from CXX to the SYCL language in
order to allow CMake to generate dependency files. This is required to
get proper dependency tracking for certain build systems, e.g. Ninja.

Unfortunately however, for GNU Makefiles the dependency tracking is
hard-coded within CMake on a per-language level. For this reason it
currently does NOT work with the SYCL language, and thus using Ninja
instead of GNU Makefiles is recommended.

This also includes a change that uses CMake's _cmake_find_compiler
function to find the SYCL compiler, instead of just assuming the
executable to be found.