clang plugin: Use custom device kernel names to avoid Clang bugs
Clang 8 and -- as of right now -- Clang 9 contain a bug that can
sometimes cause a mismatch between the names of kernels (i.e., the
__global__ entry points) within host and device binaries. This causes
the host side launch command to try and launch a kernel with a name that
does not exist within the device IR.

To avoid this issue without having to wait for a fix within Clang, this
adds a "asm label" attribute to each entry point function that is used
by the Clang name mangler, effectively setting a custom name that is
always guaranteed to be the same for both host and device side binaries.
Since the SYCL spec already includes a mechanism for specifying kernel
names (through the template parameter of parallel_for and friends), this
very mechanism is used. This means that unlike before, hipSYCL is now
sensitive to SYCL kernel names, and they can no longer be omitted.