Don't rely on OpenCL types
Currently, hipSYCL redefines OpenCL types like `cl_int` in `types.hpp`, unless OpenCL headers where already included before, in the hope of improving compatibility with SYCL applications that assume OpenCL.
In practice, this can lead to redefinition problems if the OpenCL headers are included after the SYCL headers.

This PR removes the dependency on OpenCL types for internal types used by hipSYCL, and replaces these types by custom typedefs in the form of e.g. `sycl::detail::u_int` for a 32 bit unsigned int.
This change is likely also in line with the future direction of SYCL, since it is expected that SYCL will become more independent of OpenCL.

For compatibility, the OpenCL type definitions can be reenabled with the `HIPSYCL_DEFINE_OPENCL_TYPES` macro.