Fix typo causing missing initialization for SVM buffers constructed from const pointers
In the SVM code path, `buffer_impl::write(const void* ptr)` was incorrectly copying data from the buffer's
host side buffer to to the device side buffer instead of from the argument `ptr` to the device side buffer.

Since, in SVM mode the host side buffer is the same as the device side buffer, this did effectively nothing.

This fixes this (arguably typo) to correctly copy from the argument to the device side buffer.
This also fixes the incorrect results reported for the readme's example in issue #56.