Fix issue #150: syclcc-clang complaining about missing arguments from unrelated backends
During template string substitution in cuda-link-line and rocm-link-line arguments, syclcc-clang will construct a dict() to map variable names to values.
The supported variables contain both CUDA and ROCm paths. As a consequence, these variables are accessed even if only one backend (ROCm or CUDA) is available,
causing an error.

This fixes the issue by separating the template substitution function in two functions, one for ROCm and one for CUDA.
ROCm-specific variables are now not available anymore for cuda-link-line and vice versa.