GH Action: Linux & macOS
This adds two GitHub Action scripts for CI on Ubuntu and macOS.

macOS dependencies are installed via brew, especially including
`libomp` for OpenMP support. Clang on macOS needs further a
`-Xclang` or `-Xpreprocessor` flag before its `-fopenmp` flag,
which is now added in the compiler wrapper script.

This requires CMake 3.16+ in order to query
`OpenMP_CXX_INCLUDE_DIRS`:
https://cmake.org/cmake/help/v3.16/module/FindOpenMP.html

Note that I only extract the include path - other properties such
as flags will be found for the CC/CXX compiler used to build
hipsycl and are not necessary the ones for the `clang` in
`CLANG_EXECUTABLE_PATH`.

Also avoid unknown linker flags on macOS.