Transitively add nothrow attribute to OpenMP kernel functions.
As of the SYCL spec, functions in a kernel must not throw. Therefore it is safe to mark those as nothrow (aka noexcept).
This is done in the clang frontend plugin in a similar way as the kernel functions are marked with CUDAHost and CUDADevice attributes.
This change massively simplifies the CFG that has to be adapted when performing the loop splitting as there are no invoke instructions with their respective landing pads.