Make cuda|hip_device_manager thread_local
CUDA/HIP keep track of the active device on a per-thread basis. While
the cuda|hip_device_manager classes were originally never intended to be
used from multiple threads, they in fact are (for example within the
cuda|hip_queue constructors). This can lead to subtle bugs when running
hipSYCL on a machine with multiple CUDA/HIP devices.

To remedy this, the cuda|hip_device_manager singletons are now
stored as thread_local, meaning there is a unique instance for each
new thread.