Specify alignment when allocating buffer memory.
This is required for types with required alignments, otherwise bad things might happen.
As we do currently not propagate the alignment requirements of types, I just used the element size as alignment, capped at `sizeof(double) * 16` (for `sycl::vec<double, 16>` without having to couple the rt there to libkernel)
`aligned_alloc` has implementation defined requirements regarding the alignment.
As posix & Windows require the alignment to be a power of two, I borrwoed LLVMs `PowerOf2Ceil` to make sure this requirement is met.