First (unfinished) attempts to use have memory dependencies analyzed to enable recalculation of local variables.
As a work item lives inside a loop, if we split the loop, all private state of a work item (aka loop iteration) from the loop before, is lost.
Therefore a dependency hull is generated, that contains all Instructions, that interact with a Value, that depends on the loop iteration variable. Some filtering has to be applied, as e.g. we may not overwrite global memory again. This will have to be improved.

Mostly a backup commit, so I can rebase.