[LoopSplit] Start moving towards arrayification instead of copying all depending instructions.
For each cross barrier dependency, we earlier had to copy all instructions to the new loop, which came with a bunch of difficulties, as having to determine which instructions were not overwritten yet, ..
Instead, now we keep state on stack. Thus, for each cross barrier dependency we create a 1024 elements (alloca) stack array, which we use to store the values for each iteration, which are needed again after the barrier.
This removes the need to recalculate everything, as well as the need to filter overwritten values and in theory makes handling branches much easier.