[LoopSplit] Highly experimental support for barriers in conditionals.
This first performs an analysis to find all the conditionals in the work item. Then given this information, and whether a conditional contains a barrier, this conditional is cloned and rewired, s.t. only the conditional path with the current barrier in it stays in the first loop (as we take care of one barrier at a time). The part leading to the conditional stays in the first loop. The other paths are moved into the next loop. All conditionals are cloned, but only the comparison and the branch operator.

The value dependency analysis is now moved after rewriting the loop, as we then know all blocks which are in the new and the old loops. The initial alloca arrayification stays at the beginning of the process.

DT and Loop analysis are not taken care of too much during the process right now, but using the updateDTAndLi they are still preserved.