[LoopSplit] Support more inner loops.
This adds highly experimental support for inner loops with non-constant initial loop variables and non-uniform variables. I.e. it is possible now, to have differing actual loop indices between work items (as long as the amount of iterations is the same between all work items).

This adds a bunch of todos as this is probably rather brittle and needs refactoring.