[LoopSplit] Add LLVM lit config to improve testing of the plugin.
It is now possible to use the ease of LLVM lit to test many small scenarios to check we don't break anything.
This adds some variations of a reduction testing various aspects that should work so far as well as some barriers in conditionals.

Also move instruction simplification to the end of the pass (which we probably could leave to later passes then) to allow reduce_for_inverse_barrier.cpp.