[LoopSplit] Unit tests build and run successfully!
Except for scoped parallel reduction.. but that also fails when not setting HIPSYCL_NO_FIBERS and the loops splitter should not be at work at all.

For this to work, the condition finding had to be slightly generalized and constant initialized accumulators had to be fixed.
Also the local_memory_ptr was uninitialized.