[LoopSplit] Rework handling of inner loop induction variables.
Instead of the brittle two parts used before, now the inner latch is just cleared completely and by this added to the last work-item loop with the depended upon values being stored to the corresponding alloca and loaded again in the inner latch to be available for the header's PHINodes.

Also document the process of the inner loop split a bit and slightly clean up the code.