[LoopSplit] Update nd range kernel launcher and kernel detection.
After merging current develop, the kernel launcher had to be updated to reflect the changes done for the fiber path.
This resulted in the omp parallel region to be no longer outlined in a separate function. Thus the mechanism to find kernels to split had to be adapted: the `iterate_nd_range_omp` is `clang::annotate`d with `hipsycl_nd_kernel` which can be found in the same way as the `hipsycl_splitter` annotation.
As this is also used in the frontend for the noexcept sprinkling, the dependency on using OpenMP is highly reduced.