[LoopSplit] Enable nested inner loops.
This allows the use of multiple nested loops inside of the work items with loop splitting.
This is done by first collecting all inner loops that have to be split (aka all non work item loops) and then, from the outermost to the innermost loop, repeatedly applying the loop inversing mechanism as used before. This way, the inner loop being split, is always directly inside a valid work item loop.