Don't override explicit __host__ and __device__ attributes
Previously, hipSYCL internally marked all functions called from a kernel
as `__host__ __device__`. This leads to unintended side-effects when the
code has separately defined `__host__ foo()` and `__device__ foo()`
functions.

With this patch, we only add CUDAHostAttr and CUDADeviceAttr if neither of
them is specified. If at least one is explicitly present, we respect the
user code and leave the attributes as-is.