[LoopSplit] Fix getting variable values in debugger.
The this pointer of the functor is now copied to the load block as well, to allow depending values to be resolved.
In the debug values of the inner loop induction values were not handled correctly.

Also moveInnerIndVarToWorkItemLoop is required for inner loops.

Some tests are still failing due to missing array loads in the inner loop's header.