adjusted the sub_group interface
`get_max_group_range` was deprecated, since it doesn't exist in the
specification and doesn't return the the maximum group range. It should
be removed completly in my opinion to reduce confusion.

comments added to `get_local_range` and `get_local_linear_range`, to
make it clear they always return the maximum size.

`get_max_local_range` now returns `warpSize` instead of
`get_local_range`, as the maximum is returned, and it is easy to get the
wrong impression when looking at the code and it is likely to be
fogotten when `get_local_range` is changed at some point.

This commit doesn't contain any functional changes and is purely
cosmetic.