[LoopSplit] Import VariableUniformityAnalysis and BarrierTailReplication from POCL.
This is WIP, piece by piece I aim to import POCL's kernel compiler and make it compatible with our infrastructure / use-case.

One major difference between our usecase and POCL's that has to be kept in mind, is that POCL works on a single kernel function most of the time and only in the end generates the work item loops around it. BTR seems to be fine with this, though.