[LoopSplit] Enable most regression tests with the POCL approach.
This is enabled by using a do-while like approach for the work-item loops, if the first work item was not peeled. Otherwise the for loop style is kept.
This also enables the pocl pipeline with the legacy pass manager.
Still failing are the conds_in_for, reduce_while_early_update and for_ind_cond regression tests.