[LoopSplit] Build while-switch around sub-cfgs.
This mainly replaces the original WILoop header with a block that just includes a switch branch on the LastBarrierID value mapping to the sub-cfgs.
By doing removeUnreachableBlocks, the old WILoop is completely wiped then.
Due to some context storing issues, some changes to save and restore have been necessary + adding of a last step that fixes up non-dominating dependencies in a sub-cfg.

The compiled program runs successful in debug, but has some stack access issues. This will be handled in a later commit.