[LoopSplit] Make custom alloca arrayification for CBS.
This eliminates unnecessary arrayification for allocas that are only used in one subcfg. For this a hull of all users that may read/write or work on a pointer to the alloca is created and checked whether it's in a single cfg.
Todo: construct a test case where this might be the case for two subcfgs but actually having a dependency..