[LoopSplit][CBS] Single-SubCFG values cannot blindly only be loaded in the load block.
If the value in question is e.g. used as induction variable in an inner loop, it is not sufficient to load it once per WI, but instead is has to be reloaded at every loop iteration -> load has to happen immediately before the first use.

Also add test for the for{if} edge case that is only supported by csb.