[LoopSplit] Fix marking predeclared kernel functions noexcept.
The CCS does not create the full hull of called functions if the functions are predeclared, as they might not yet be defined when VisitFunctionDecl is called. Thus, the actual noexcept marking has to be done when the translation unit is complete, aka in `HandleTranslationUnit`.

Also add the correct tag for the inlining log messages and replace barrier functions with __hipsycl_barrier for simplicity.