Workaround CUDA / MS ABI mismatch on Windows.
On Windows the MS ABI is used for host code, that also influences the outline of structs.
For some obscure reason multiple empty base classes lead to 8 byte padding in MS ABI.
CUDA device code on the otherhand uses Itanium ABI.
Itanium ABI doesn't randomly add padding for empty base classes.
As the size of the accessor variants thus differs on host and device, invalid accesses occur, as the device loads the pointer from the kernel lambda from the wrong position -> bumms.
The standard SYCL accessors are fine as they don't have any empty base classes.

To mitigate this, we add __declspec(empty_bases) as presented in https://devblogs.microsoft.com/cppblog/optimizing-the-layout-of-empty-base-classes-in-vs2015-update-2-3/ to get the Itanium compatible layout.