Ensure unique kernel names with attributes in HIP/CUDA.
Adding the property list to the kernel name ensures that for e.g. every unique reqd_work_group_size, a unique kernel is generated. With this in place, compiler hints may be generated from the property lists without invalidly mixing kernels and attributes.

(Note: SPIR-V uses effective_type in conjunction with __builtin_unique_stable_name, which will have to be tested and given its removal and refactoring in Clang > 11, will have to be looked at anyways..)