[CBS] Use UniAna for contiguous values.
The idea behind this, is that e.g. the global ids in theory are contiguous values but currently are stored inside an arrayified alloca, which makes it impossible for e.g. the vectorizer to detect that loads to global memory are not random access but indeed contiguous.

To actually greatly benefit from this, we have to be able to identify the global id as contiguous. This does currently not happen, as the id.linear_id stuff leaves a loop at the stage we use it, with the global id as PHI node.
As the initial value stored to the ex_phi alloca is 0, it's first detected as uniform but the following values are contiguous. The VectorShape::join correctly identifies the alloca as varying then.