[LoopSplit] Compat with LLVM<11
Dump plugin information to syclcc.json

With this information, the heuristic, whether "omp" shall be mapped to the accelerated variant doesn't stumble when a too old LLVM version is used.
Also use the information whether the cpu acceleration was built instead of the OS.

Also don't build the CPU acceleration support with too old clang.