[CBS] Dynamically sized stack arrays.
Backporting fixes done for upstreaming CBS to POCL.
This is mostly using an llvm::Value as alloca size for the local stack arrays -> in theory we don't have to limit the CPU work-group size to 1024 anymore.
This also allows for larger work-item private arrays, if the actual work-group size is < 1024.