Fix computation of `end` iterator
Since we're computing everything w.r.t. to the offset
and within the sub-range, the end-iterator is just the
product of the values of each dimension of the sub-range.
If the accessor is unranged, the range to consider is just
the whole range of the underlying buffer. In any case, this
is given by `accessor.get_range().size()`.