Remove declaration of `ret` to remove need for default constructor of `return_type`
When the macro `HIPSYCL_STDPAR_OFFLOAD` is used in `std::for_each_n`,
the type `return_type` is `ForwardIt` (in the C++ standard it is
called `InputIt`) and this iterator only has to meet the requirements
of a `LegacyInputIterator` which does not include default-constructibility.
By replacing the if/else with a ternary operator, the declaration that
tries to call the default constructor can be removed.