[CMake] Add `ACPP_HAS_*` values to CMake integration.
So far, to my knowledge it is not easily possible to query on the CMake level whether the used AdaptiveCpp was built with certain features.
This adds a bunch of feature flags to the CMake integration making this quite simple.

E.g. CMake users can now just switch on `ACPP_HAS_SCPP` which target to set, `omp;generic` if SSCP is available, `omp;cuda:sm_80,sm_75;hip:gfx90c` or sth like this otherwise (the backend flags might help determine which targets to actually include ;-)).