[SSCP][Host] Use distinct library file to not rely on persistent kernel cache.
If `ACPP_RT_NO_JIT_CACHE_POPULATION` the kernel cache is not filled. Therefore, the host SSCP must not rely on the cache being populated.
With this patch, the omp_queue writes their own library file upon JIT time. This is cleaned up after the run. Currently this also lives in the JIT cache folder. We could also move that to `/tmp`, which potentially has performance benefits as it usually is a tmpfs..