[CBS] Get way more conservative on when not to widen allocas.
We have to widen allocas, so that each work item get's it's own alloca slot, whenever, the alloca is used across multiple SubCFGs.
This commit adds into consideration that this might as well happen, when the pointer to the alloca is stored somewhere else.
This can happen whenever the pointer is stored or if it is being used by a function.
When it's used by a function, we try to be nice and don't arrayify whenever the pointer operand is marked as "nocapture" or as "sret" (struct return, i.e. being initalized).