[CBS] Insert PHIs manually if required.
When splitting a loop in CBS, we might get "instruction not dominated" issues.
Most of them were already captured by the existing infrastructure.

However, the issue with #1898, was that an inner loop was split, and the outer loop wasn't split.
All loop indices are uniform in the outer loop.
Therefore, they were loaded only once in the uniload block.
However, when the outer loop takes its backedge (after the inner loop exited, thus without another barrier), the outer loop's index is not updated for the first iteration of the inner loop.
Therefore, we should create a PHI node after the backedge of the outer loop to select between the value loaded in the uniload block and the value from the backedge of the outer loop.