[CBS] Use dominance frontier to find all the spots where a PHI might be needed.
A little intro on DFs and why they're interesing for this task:
https://pages.cs.wisc.edu/~fischer/cs701.f05/lectures/Lecture25.4up.pdf

This is a more fundamental approach to the issue, so I hope it'll hold up against whatever we may throw against it.

At the moment, this is an overapproximation, i.e. there might be potential to prevent creating too many PHIs, but I expect them to be optimized away, as they're just dead code.