Use aspect selector for host USM SYCL test
Rather than using the GPU device selector to enable a SYCL
test that requires host USM, use a aspect selector for
the host usm capability.

This change allows the test to run on USM enabled devices that
aren't GPUs, in particular OpenCL CPU implementations with the
Intel USM extension feature.